Refactor assessments API and add tests
Rename assessment helper modules and query/mutation files to
standardize names (questionBanks → bank, richContent → content,
assessmentList → list, saveResponse/startAttempt/submitAttempt →
save/start/submit) and update imports/usages across the frontend and
generated API types. Add seed helpers and focused Vitest suites for
create, delete, update, and submit flows. Remove the obsolete
"assessment:archive" permission and related i18n keys.
