+++ apps/www/components/school/classes/assessments/types.ts
@@ -3,5 +3,5 @@ import type { FunctionReturnType } from "convex/server";
 
 /** One authored assessment row returned by the class assessments paginated query. */
 export type Assessment = FunctionReturnType<
-  typeof api.assessments.queries.public.assessmentList.listAssessments
+  typeof api.assessments.queries.public.list.listAssessments
 >["page"][number];
