+++ packages/backend/convex/assessments/helpers/bank.ts
