+++ packages/backend/convex/assessments/mutations/public/create.test.ts
@@ -1,0 +1,64 @@
+import { api } from "@repo/backend/convex/_generated/api";
+import {
+  insertClass,
+  insertSchool,
+  NOW,
+  PARAGRAPH,
+} from "@repo/backend/convex/assessments/seed";
+import {
+  createConvexTestWithBetterAuth,
+  seedAuthenticatedUser,
+} from "@repo/backend/convex/test.helpers";
+import { describe, expect, it, vi } from "vitest";
+
+describe("assessments/mutations/public/create", () => {
+  it("creates an authored assessment with a derived slug and order", async () => {
+    vi.setSystemTime(new Date(NOW));
+
+    const t = createConvexTestWithBetterAuth();
+    const seeded = await t.mutation(async (ctx) => {
+      const teacher = await seedAuthenticatedUser(ctx, {
+        now: NOW,
+        suffix: "create-teacher",
+      });
+      const schoolId = await insertSchool(ctx, teacher.userId);
+      const classId = await insertClass(ctx, schoolId, teacher.userId);
+
+      await ctx.db.insert("schoolMembers", {
+        schoolId,
+        userId: teacher.userId,
+        role: "admin",
+        status: "active",
+        joinedAt: NOW,
+        updatedAt: NOW,
+      });
+
+      return { classId, schoolId, teacher };
+    });
+
+    const teacherClient = t.withIdentity({
+      subject: seeded.teacher.authUserId,
+      sessionId: seeded.teacher.sessionId,
+    });
+
+    const assessmentId = await teacherClient.mutation(
+      api.assessments.mutations.public.create.createAssessment,
+      {
+        schoolId: seeded.schoolId,
+        classId: seeded.classId,
+        title: "Assessment 1",
+        description: PARAGRAPH,
+        mode: "assignment",
+        status: "draft",
+      }
+    );
+
+    const assessment = await t.query(async (ctx) => {
+      return await ctx.db.get("schoolAssessments", assessmentId);
+    });
+
+    expect(assessment?.slug).toBe("assessment-1");
+    expect(assessment?.order).toBe(0);
+    expect(assessment?.status).toBe("draft");
+  });
+});
