+++ packages/backend/convex/assessments/mutations/public/questions.ts
@@ -2,7 +2,7 @@ import {
   requireAssessment,
   requireAssessmentPermission,
 } from "@repo/backend/convex/assessments/helpers/access";
-import { requireRichContentSize } from "@repo/backend/convex/assessments/helpers/richContent";
+import { requireRichContentSize } from "@repo/backend/convex/assessments/helpers/content";
 import { richContentValidator } from "@repo/backend/convex/assessments/schema";
 import { mutation } from "@repo/backend/convex/functions";
 import { requireAuth } from "@repo/backend/convex/lib/helpers/auth";
