+++ packages/backend/convex/assessments/mutations/public/update.ts
@@ -3,8 +3,8 @@ import {
   requireAssessment,
   requireAssessmentPermission,
 } from "@repo/backend/convex/assessments/helpers/access";
+import { requireRichContentSize } from "@repo/backend/convex/assessments/helpers/content";
 import { validateScheduledStatus } from "@repo/backend/convex/assessments/helpers/publishing";
-import { requireRichContentSize } from "@repo/backend/convex/assessments/helpers/richContent";
 import { richContentValidator } from "@repo/backend/convex/assessments/schema";
 import { mutation } from "@repo/backend/convex/functions";
 import { requireAuth } from "@repo/backend/convex/lib/helpers/auth";
