+++ packages/backend/convex/assessments/seed.ts
@@ -1,0 +1,49 @@
+import type { Id } from "@repo/backend/convex/_generated/dataModel";
+import type { MutationCtx } from "@repo/backend/convex/_generated/server";
+
+export const NOW = Date.UTC(2026, 3, 16, 14, 0, 0);
+
+export const PARAGRAPH = {
+  format: "plate-v1" as const,
+  json: JSON.stringify([{ type: "p", children: [{ text: "Hello" }] }]),
+  text: "Hello",
+};
+
+/** Insert one school row with the minimum fields required by the schema. */
+export async function insertSchool(ctx: MutationCtx, userId: Id<"users">) {
+  return await ctx.db.insert("schools", {
+    name: "Nakafa School",
+    slug: `nakafa-${userId}`,
+    email: `${userId}@example.com`,
+    city: "Jakarta",
+    province: "DKI Jakarta",
+    type: "high-school",
+    currentStudents: 0,
+    currentTeachers: 0,
+    updatedAt: NOW,
+    createdBy: userId,
+    updatedBy: userId,
+  });
+}
+
+/** Insert one class row with the minimum fields required by the schema. */
+export async function insertClass(
+  ctx: MutationCtx,
+  schoolId: Id<"schools">,
+  userId: Id<"users">
+) {
+  return await ctx.db.insert("schoolClasses", {
+    schoolId,
+    name: "Class 10A",
+    subject: "Mathematics",
+    year: "2026/2027",
+    image: "retro",
+    isArchived: false,
+    visibility: "public",
+    studentCount: 0,
+    teacherCount: 0,
+    updatedAt: NOW,
+    createdBy: userId,
+    updatedBy: userId,
+  });
+}
