+++ packages/internationalization/dictionaries/en.json
@@ -736,10 +736,8 @@
       "create-assessment-description": "Add an assessment to this class.",
       "edit-assessment-title": "Edit assessment",
       "edit-assessment-description": "Update the details of this assessment.",
-      "archive-assessment": "Archive",
       "assessment-created": "Assessment created.",
       "assessment-updated": "Assessment updated.",
-      "assessment-archived": "Assessment archived.",
       "assessment-deleted": "Assessment deleted.",
       "create-assessment-failed": "Couldn't create the assessment. Please try again.",
       "update-assessment-failed": "Couldn't update the assessment. Please try again.",
