+++ packages/internationalization/dictionaries/id.json
@@ -736,10 +736,8 @@
       "create-assessment-description": "Tambahkan asesmen untuk kelas ini.",
       "edit-assessment-title": "Ubah asesmen",
       "edit-assessment-description": "Perbarui detail asesmen ini.",
-      "archive-assessment": "Arsipkan",
       "assessment-created": "Asesmen berhasil dibuat.",
       "assessment-updated": "Asesmen berhasil diperbarui.",
-      "assessment-archived": "Asesmen berhasil diarsipkan.",
       "assessment-deleted": "Asesmen berhasil dihapus.",
       "create-assessment-failed": "Gagal membuat asesmen. Silakan coba lagi.",
       "update-assessment-failed": "Gagal memperbarui asesmen. Silakan coba lagi.",
